Adding own math library to substitute GLM
First files implemented with all the functions necessary for the engine. Added vec2, vec3, vec4 and mat4

+++ Ilargi/source/Platform/Vulkan/VulkanMaterial.cpp
@@ -32,7 +32,7 @@ namespace Ilargi
 		descriptorWrites[0].descriptorCount = 1;
 		descriptorWrites[0].pImageInfo = &imageInfo;
 
-		vkUpdateDescriptorSets(device, descriptorWrites.size(), descriptorWrites.data(), 0, nullptr);
+		vkUpdateDescriptorSets(device, static_cast<uint32_t>(descriptorWrites.size()), descriptorWrites.data(), 0, nullptr);
 	}
 	
 	VulkanMaterial::~VulkanMaterial()
